:art: cleanup and factor out fancyElementFactory for better coding
